Move AZ funtions code
